Load and Save functionality!
Added two new classes, SaveFrame and LoadFrame. These are used to create a frame where the user can write the name of the file they want to
save and the name of the file they want to load.
Added a little more comments, both to the SaveFrame and LoadFrame but also the ActionBar.