Selection List and Component Name
Added a JList based on a string array with name of the components.
The size of the array should match how many components that is avaiable.

Added text in the paint function for the GridNode to display what kind of component that is placed.
This will help seperate the components with similar or the same graphic (for example, OR and AND gates).

NOTE; old dat files may not work or may.